billing/collector: rename internal package to pkg.
