+++ systems/billing/collector/pkg/global.go
@@ -1,4 +1,4 @@
-package internal
+package pkg
 
 const ServiceName = "collector"
 const SystemName = "billing"
